ai-wiki-quiz-main/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
beautifulsoup4==4.12.3
requests==2.31.0
pydantic==2.5.3
langchain-core==0.1.23
langchain-community==0.0.20
langchain-google-genai==0.0.6
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pymysql==1.1.0

